Let the film-ratings route declare its own trust boundary
CI caught the auth gate being invisible on the route: it was applied at
the include_router call because get_active_upload_user lived in main,
which a route module cannot import (main imports the routers first, so
the name does not exist yet). Anything inspecting the route saw only
get_upload_user, and one test could not resolve the route at all.

Moving the dependency into api/dependencies.py removes the cycle, so
the route declares its gate directly and the mount carries none.
Verified by calling it: 401 unauthenticated and 401 with a bad token.
